[OVN] Do not supply gateway_port if it's not binded to chassis
There is no reason to define gateway_port explicitly while create NAT
rules for FIPs in case where LRP is not binded to chassis. Doing so
will result in broken flows making NAT rules effectivelly ignored.

This is a case, when Logical_Router is pinned to chassis instead of LRP

Closes-Bug: #2083527
Change-Id: Iafaed72b6ac295db88497e16c54090952c410194